Machine Learning/requirements.txt
# Credit Card Fraud Detection — ML Phase
# =========================================
# Core ML
scikit-learn>=1.4
imbalanced-learn>=0.12
pandas>=2.1
numpy>=1.26
joblib>=1.3

# Hyperparameter Tuning
optuna>=3.5

# Experiment Tracking (optional)
mlflow>=2.10

# Visualisation
matplotlib>=3.8
seaborn>=0.13

# Database (optional — for Power BI integration)
psycopg2-binary>=2.9

# Notebook support
jupyter>=1.0
ipykernel>=6.0
